+++ HPL2/core/include/input/Action.h
@@ -119,7 +119,7 @@ namespace hpl {
 		void Update(float afTimeStep);
 
 		iSubAction *GetSubAction(int alIdx){ return mvSubActions[alIdx];}
-		int GetSubActionNum(){ return (int)mvSubActions.size();}
+		size_t GetSubActionNum(){ return mvSubActions.size();}
 
 
 		const tString& GetName(){ return msName;}
